Add methods to set and retrieve format indices for a blend value node
In order to have a consistent data layout for all interior nodes of a
blend tree, we need a place to convert from whatever data layout a
clip produces when evaluated to the format used  by the blend tree. We
will do this by creating for each value leaf node a vector of indices
that can be used to map from the ClipResults layout output by the clip
to the ClipResults format used by the blend tree.

With this approach, we can do the following:

* Evaluate each leaf node in the blend tree
* Perform a gather operation using these format indices to reorder the
  data into a consistent layout for the whole blend tree
* Evaluate the blend tree using this consistent layout

This completely avoids having to perform complex mappings when
evaluating each blend node, which is a significant saving in
overhead and allows for more parallellism in the future.

Change-Id: Ie534534f9555842d0d0f1a89fc996e25f5c6ce9e
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>