Add function assignChannelComponentIndices
This iterates through the channel description and assigns channel
component indices in ascending order.

The next step will be to use this in conjunction with the channel
description with the layout of channels in each clip to generate
the format indices that will allow us to quickly map from the clip's
raw ClipResults to ClipResults formatted in the layotu used by the
blend tree/animator.

Change-Id: I04dd5a21ca8355529118cfb2b05f4904e43ef0e2
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>