Don't explicitly disable the Qt3D category logging
Category logging is off by default, so explicitly turning it off will
just make it harder to turn it on in the application.

qt.*.debug has special behavior in the sense that defaults to off. Other
categories, like Qt3D.*.debug default to true. Therefore the only correct
way in Qt 3D is to always specify warning-and-up as the default enabled
severity.

Change-Id: I91daf48e049b7a9fcfac9720982138b178ee4804
Reviewed-by: Antti Määttä <antti.maatta@qt.io>