Handle multiple surfaces properly for picking
This fixes issues with picking when having multiple viewports and/or
multiple windows.

Now check that the mouse event actually hits inside the viewport extents.

Also track the source of events and check that the surface matches the
source of the event.

Remaining issue is overlapping viewports within the same window

Task-number: QTBUG-59567
Change-Id: I76a4ee2bec7300d893fef6040d89bf81f2109795
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>