Add download service and use for remote meshes
Add a new QDownloadHelpService. Can be provided with
QDownloadRequests which will trigger downloads and return
the data in a QByteArray.

Downloads are triggered in a separate thread and requests
have the opportunity to do long running operations there
(it will block other requests though).

Implemented for downloading meshes. When the functor is
triggered, it checks the url and submits a request for
download. When that completes, it’ll save the data in the
functor and mark the component as dirty that that a job
will run again to complete the parsing.

LATER (5.10):
- status property, progress maybe

Task-number: QTBUG-57614
Change-Id: I05abe9610e12670a368d480b2fb3115987201e6b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>