Renderer: fix OnDemand rendering
Note: this is a risky change

For some reason the renderer assumed that a RenderQueue would never have a
size of 0. This prevented the renderer from calling proceedToNextFrame() on
the vsync advance service as it somehow felt that it was a case of rendering
with a RenderQueue not yet ready. In the case of OnDemand rendering it is
perfectly valid to have a RenderQueue of 0 (nothing has changed that requires
rendering) but we still need to call proceedToNextFrame() otherwise
syncChanges() is never called and we end up never updating the aspects ever
again.

The renderer was updated to
1) check if a RenderQueue is complete
2) check if the RenderQueue is empty.
(a RenderQueue can be complete but empty (OnDemand case))
3) Proceed to next frame if RenderQueue is complete or RenderQueue is empty.

Change-Id: I27ae778831c9b136db1e1a69892f6fde291fd965
Task-number: QTBUG-59696
Task-number: QTBUG-54900
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>