AssimpImporter: properly handle texture wrapping
By default the wrap mode is Repeat. If specified it is changed to the
appropriate value.
Note: This changes also shows that each material could be definining the same
texture with different wrap modes. Therefore we cannot store textures only
based on their file path like we used to.

Change-Id: Ic543c466d7c5cc7cd7e26d1738006952a8ce746b
Task-number: QTBUG-59531
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>