Pick the right type for QVariant channel targets
Animating a Parameter node would be impossible otherwise since
the 'value' property is a QVariant so QChannelMapping has no chance
of know what really is expected there.

If a value is already set, we can automatically determine the QVariant's
underlying type. If not, show a warning since we have no chance then.

Start handling QMetaType::Float as well since a QVariant will ofen
contain this, not a Double.

Change-Id: I0ba2abbb2c8d85f2aa1a152ed4b8c6cd26f1fc97
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>