Support setting texture format on QTextureLoader
This allows to specify the data loaded from a png image should be
using the SRGB8_Alpha format instead of the usual RGBA8_UNorm format
such that we get hardware sRGB->linear color conversion when sampling
from such a texture. This is needed for serious use of PBR as the
albedo/base color texture is often sRGB encoded.

Task-number: QTBUG-60977
Change-Id: I0fd2baf533bb69eaf9e4ae5b696e2841f89a51e7
Reviewed-by: Robert Brock <robert.brock@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>