Fix context sharing in scene2d
The scene2d context sharing fails if the share context is in use when
scene2d tries to create it's context. This happens when the scene2d is
initialized while Qt 3D is already using it's own context. In Scene3D case
we always create another context to prevent this. Change the Qt 3D to
create it's own share context only when it's context shareContext is not set.

Also change the format in scene2d to always get it from the share context
so that it matches with it and the sharing works with all platforms.

Task-number: QTBUG-60974
Change-Id: I11df0b647207dbcc1d285ae4687c1e37010443f2
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>