Fix componentAdded/componentRemoved notifications dispatch to components
When a component is added to an entity, two change
notifications were sent but they both were directed
to the entity.
This makes sure one message is sent to the entity
and one to the component, as was intended.

Same for remove notifications.

Fixed the unit tests which where incorrectly testing for the
subject id of the change notification.

Change-Id: Ic697c08e92e884abf28e675e4ae2b44a93ca45d5
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>