Port Metal/Rough materials to shader graphs
This allows to validate the approach with a couple of functioning
materials using the new shader generation system. Also allows to have a
few prototypes in the default set of nodes.

Clearly the system is still cumbersome to use as can be seen from the
multiplication of prototypes, we shall improve that next by allowing
more expressivity in the graph JSON to allow more control (clearly
telling on the uniforms, return type for the functions, etc.).

Change-Id: I3edfef65af288faa339661f20d5e640cf00e9234
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>