Add support for channel components with no name
Some API clients are able to generate properly ordered channel component
lists on their own. (properly ordered meaning X=0, Y=1, etc.)

It is a waste of time to perform string comparisons in this case so just
recognize this case (the QChannelComponent name is simply unset, empty)
and use the index as-is.

Change-Id: I7aae5f3be427e60e705ef3dd174c0bf5d1fdee19
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>