Add QSkeleton and QSkeletonLoader
They both inherit from QAbstractSkeleton which exists purely to
provide a type for setting skeletons on the upcoming QArmature
component. The classes relate in the same way as
QAbstractAnimationClip, QAnimationClip and QAnimationClipLoader.

The skeleton provides the joint hierarchy that can be animated and
then used to skin a mesh via the upcoming QArmature component.

Change-Id: I4be8ce3f1d6cbca4af08c5829b33172a2f951b11
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>