Make GLTexture dirty flags atomic instead of using a mutex
The mutex right now is used for controlling access to the m_dirty flags,
and for guarding access to m_gl and other data.

We don't need to use it for accessing the flags, as an atomic int will
do just fine and relieve us of potential deadlocks. The mutex must only
guard changes to the actual data.

Task-number: QTBUG-61130
Change-Id: Ia1f25af2233387f375c077965e901c67f972f1ec
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>