Only create the functor if the engine is available
Otherwise, when it becomes available via the scene, we end up setting
the functor again leading to repeated work. So defer setting the
functor until the engine is available to just do the work once.

Also fix up operator == to check mesh name as well as url.

Updated unit tests to reflect.

Change-Id: I41d6e0670a6b1f582899953cf858e842b6a214ff
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>