Use layers to have only one graph for Metal/Rough
Both QMetalRoughMaterial and QTexturedMetalRoughMaterial use the same
graph now. Also QMetalRoughMaterial does the job to be clever and
activate the right layers depending on the type set on the properties
(which is why they are now all variants).

QMetalRoughMaterial can then be used as a drop in replacement for
QTexturedMetalRoughMaterial.

Change-Id: Ia52fe4608aca5558642599ed435f3cd9ac93e651
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>