Fix StringToInt implementation
Reduces the overhead per call of lookupId() from around 7000 to 180
instructions for a use case. Should in general scale much better.

I still do not like that this is using global static data that is
never cleaned up. It should be tied at the minimum to the lifetime
of the Qt 3D scene.

Change-Id: I777a2bbf2d765f4b0fc9e3d2b06692c7260f5a9f
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>