Fix race in renderer
All buffers are uploaded to GPU before the frame submission takes place.
If during frame submission we check for a CPU buffer whether it's dirty, we end up
(rightfully) with a race condition as we are preparing for frame n+1 already.
There must have been a left over when the above changes were made as the
UBO/SSBO code paths would, during rendering, check whether a CPU buffer was
dirty, and if so try to upload, causing a race.
Removed these code paths which aren't needed.

Change-Id: I713a92b881335ecfb30f24bd7485a7bac29be4b7
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>