Don't clear the target indices vector when building the format
This will impact on the future format building when there are
multiple clips to be blended. For now we fill them with -1 values and
alter the check when building the channel mapping data structures.

A follow up commit will make the input to this function immutable and
instead output the masked off channels as part of the format which can
be used to build the channel mappings.

Change-Id: I7eb42041ccfe253a449f13fde89236eab56dba18
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>