Render next frame if texture data is not available
Texture data from generators may not be available in the current frame and
the next frame should be rendered immediately, even in OnDemand rendering.
The problem was that this information was not propagated back to the
renderer, it was just silently ignored inside
GraphicsContext::setParameters. Added a bool return type to
GraphicsContext::setParameters so that the renderer could check for
errors.

Task-number: QTBUG-63561
Task-number: QTBUG-59752
Change-Id: I18fb6f8236bf47d3eeab07d879dffaa599665e6b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>