Fix material system qrc in static builds
The Qt3DRender part is clear and we have a good global place to do the
Q_INIT_RESOURCE call.

Extras is debatable since there is no good place. Quick is already handled
via the import plugins, but pure C++ applications also need this in static
builds. For now add it to QForwardRenderer and Qt3DWindow. This should
cover the common use cases for the materials in extras.

Task-number: QTBUG-64279
Change-Id: I276e8d4265f4527736d415b07dbf03e9d8904adf
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>