Scene3DRenderer: ensure window pointer stays the same while rendering
It appears that the pointer value for the window could in some cases be
updated meanwhile we were rendering, which resulted in random crashes.

We now save the window pointer to a member variable on startup or whenever the
windowChanged signal is fired and use a mutex to ensure the pointer won't
change while we are rendering.

This also covers the case where the item is destroyed but the cross-thread
signals are not received before rendering happens.

Task-number: QTBUG-63897
Task-number: QTBUG-65407
Change-Id: I5f2797e2210b532f9086ed186959fce27ea9f514
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>