Scene3DRenderer: ensure window pointer stays the same while rendering
It appears that the pointer value for the window could in some cases be
updated meanwhile we were rendering, which resulted in random crashes.

We now save the window pointer to a member variable on startup or whenever the
windowChanged signal is fired and use a mutex to ensure the pointer won't
change while we are rendering.

Change-Id: I93ade6bd549886b75140484c83fe59187d47ed9f
Task-number: QTBUG-63897
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>