Move animation callback handling to QAbstractClipAnimator
The code that triggers the callbacks on the GUI thread should be the used
for all clip animators, and not just the QClipAnimator.
This change also fixes the scene change event handler to call the parent
implementation for all events not handled by the clip animator, so it
doesn't silently drops all events not handled explicitly.

Change-Id: I0b1f57f362f2dc2e82e05d432426cc46d4a100ae
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>