tst_QFrameAllocator: actually test something
The old code created 30 'composed' objects, but failed to add them
to the 'composedValues' container, over which it iterated later to
perform some checks. Since the objects were never added in the first
place, nothing was actually checked.

Fix a wrong expected value, now that the test is actually testing
something.

Change-Id: I69a91af1f609ad1b6fa147aa6c5733a6a91c7ca0
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>