Remove hardcoded C++ version for MinGW/GCC
This was apparently needed for older MinGW-w64 headers that did not
properly define some common system functions otherwise. Anyhow, it is
not required anymore for newer MinGW-w64, and actually breaks
compilation with gcc 6 and newer: There we enable shift-overflow=2,
which also warns about shifting into the sign bit for C++11, but not
for C++14.

Task-number: QTBUG-60249
Change-Id: I53bc858c6f8baf9b60940904b3b28f7a49ab28e6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>