Calculate and store real transform of camera in addition to view matrix
We used to store the viewMatrix as returned by QMatrix4x4::lookAt
in the transform component of the QCamera.
This was unfortunate for two reasons:

1) An arbitrary entity could not be used as a camera without changing
   its transform to a viewMatrix, which breaks any other use of the
   entity.
2) Adding entities as children to cameras lead to the wrong
   transformation. This made it impossible to properly add objects
   that should be attached to the camera, such as the hands of a
   character in first-person view.

This commit computes the transform of the camera based on the
requested position, view center and up vector.
The view matrix is calculated when the matrices are updated in
the RenderView. This calculation assumes that the view direction
is along the entity's negative z-axis and that the up vector
is along the y-axis.

Change-Id: If22b29e3d38bf55fbf79e9f7baf0c922e0a48aeb
Reviewed-by: Mike Krus <mike.krus@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>