Deprecate viewMatrix() on QCamera because its name is misleading
The matrix returned by viewMatrix() is in a coordinate system relative
to its parent, which is not the final view matrix used when rendering.
The final view matrix depends on the world transform of the camera,
but we do not calculate this on the frontend for performance reasons.
We therefore deprecate this function because its name is misleading.

Change-Id: Ib7891c13d83b4de43b86dde878994f66979ecb02
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>